requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==2.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
betterpath==0.2.2
black==22.12.0
bleach==5.0.1
blessed==1.19.1
blinker==1.5
boto3==1.26.31
botocore==1.29.31
bpython==0.23
Brlapi==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
cliapp==1.20180812.1
click==8.1.3
click-config-file==0.6.0
cmdtest==0.32+git
colorama==0.4.6
comm==0.1.2
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
cssselect2==0.7.0
cupshelpers==1.0
curlify==2.2.1
curtsies==0.4.1
cwcwidth==0.1.8
dbus-python==1.2.16
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
define==1.74
defusedxml==0.7.1
dia==0.1.8
diff-match-patch==20200713
distlib==0.3.6
distro==1.8.0
distro-info==1.0
Django==4.1.4
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-credit-cards==0.4.1
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.0.2
django-js-asset==2.0.0
django-phonenumber-field==7.0.1
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-shortcuts==1.6
django-storages==1.13.1
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.4
duplicity==0.8.12.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
facebook-business==15.0.0
facebook-sdk==3.1.0
fasteners==0.18
fastjsonschema==2.16.2
feedparser==6.0.10
filelock==3.8.2
flake8==6.0.0
Flask==2.2.2
flask-swagger==0.2.14
Flask-WTF==1.0.1
fqdn==1.5.1
freezegun==1.2.2
frozenlist==1.3.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
gunicorn==20.1.0
html2text==2020.1.16
html5lib==1.1
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-server==1.23.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server_terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
keyring==23.11.0
l==0.11.0
language-selector==0.1
launchpadlib==1.10.18
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libsass==0.22.0
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
maven==0.0.6
mccabe==0.7.0
mistune==2.0.4
mock==4.0.3
monotonic==1.6
more-itertools==9.0.0
multidict==6.0.3
mypy==0.991
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
ng==1.2.0
notebook==6.5.2
notebook_shim==0.2.2
num2words==0.5.12
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
odoo==16.0.post20221124
odooku==11.0.14
odooku-odoo-account-invoicing==11.0.7
odooku-odoo-addons==11.0.7
odooku-odoo-base==11.0.7
odooku-odoo-board==11.0.7
odooku-odoo-calendar==11.0.7
odooku-odoo-contacts==11.0.7
odooku-odoo-crm==11.0.7
odooku-odoo-extra==11.0.7
odooku-odoo-fleet==11.0.7
odooku-odoo-hr==11.0.7
odooku-odoo-hr-attendance==11.0.7
odooku-odoo-hr-expense==11.0.7
odooku-odoo-hr-holidays==11.0.7
odooku-odoo-hr-recruitment==11.0.7
odooku-odoo-hr-timesheet==11.0.7
odooku-odoo-im-livechat==11.0.7
odooku-odoo-lunch==11.0.7
odooku-odoo-mail==11.0.7
odooku-odoo-maintenance==11.0.7
odooku-odoo-mass-mailing==11.0.7
odooku-odoo-mrp==11.0.7
odooku-odoo-mrp-repair==11.0.7
odooku-odoo-note==11.0.7
odooku-odoo-point-of-sale==11.0.7
odooku-odoo-project==11.0.7
odooku-odoo-purchase==11.0.7
odooku-odoo-sale-management==11.0.7
odooku-odoo-stock==11.0.7
odooku-odoo-survey==11.0.7
odooku-odoo-website==11.0.7
odooku-odoo-website-blog==11.0.7
odooku-odoo-website-event==11.0.7
odooku-odoo-website-forum==11.0.7
odooku-odoo-website-sale==11.0.7
odooku-odoo-website-slides==11.0.7
ofxparse==0.21
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.10
oscrypto==1.3.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
passlib==1.7.4
pathspec==0.10.3
pexpect==4.8.0
phonenumbers==8.13.2
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2022.11.30
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
pluggy==1.0.0
polib==1.1.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycogreen==1.0.2
# psycopg2==2.8.4
# psycopg2-binary==2.9.5
ptyprocess==0.7.0
pur==7.0.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
pycups==1.9.73
pydot==1.4.2
pyflakes==3.0.1
Pygments==2.13.0
PyGObject==3.36.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.7
pyinotify==0.9.6
PyJWT==2.6.0
pyldap==3.0.0.post1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF2==2.12.1
PyPDF3==1.0.6
pyRFC3339==1.1
pyrsistent==0.19.2
pyserial==3.5
pytest==7.2.0
python-apt==2.0.0+ubuntu0.20.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.49
python-dotenv==0.21.0
python-json-logger==2.0.4
python-ldap==3.4.3
python-stdnum==1.18
python-twitter==3.5
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyusb==1.2.1
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.3.1
redis==4.4.0
reportlab==3.6.12
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests-unixsocket==0.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rwanda==0.2.1
s3transfer==0.6.0
saturn-proxy-server==0.9
scikit-learn==1.2.0
scipy==1.9.3
SecretStorage==3.3.3
Send2Trash==1.8.0
sgmllib3k==1.0.0
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlparse==0.4.3
ssh-import-id==5.11
stack-data==0.6.2
suds-jurko==0.6
svglib==1.4.1
systemd-python==234
tablib==3.3.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
ttystatus==0.38
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uri-template==1.2.0
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
usb-creator==0.3.7
vatnumber==1.2
vcversioner==2.16.0.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
vobject==0.9.6.1
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
whitenoise==6.2.0
WTForms==3.0.1
xhtml2pdf==0.2.8
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.8.2
zeep==4.2.1
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
